+++ components/logframe/output-indicators-table-columns.tsx
